Moved TestFQDNCacheMinTTL to antreapolicy_test.go and modified WithCustomDNSConfig. - moved the FQDNCacheMinTTL test to antreapolicy_test file where other related tests are present   and deleted the standalone file that had earlier existsed. - modified WithCustomDNSConfig function to accept DNS config as a parameter of type *corev1.PodDNSConfig.
